Extract tests to their own files
