+++ app/gui2/src/components/DocumentationPanel/history.ts
@@ -48,50 +48,3 @@ export class HistoryStack {
     return this.index.value < this.stack.length - 1
   }
 }
-
-if (import.meta.vitest) {
-  const { test, expect } = import.meta.vitest
-  const ID_1 = 10
-  const ID_2 = 20
-  const ID_3 = 30
-
-  test('HistoryStack basic operations', () => {
-    const stack = new HistoryStack()
-    expect(stack.current.value).toBeUndefined()
-    expect(stack.canGoBackward()).toBeFalsy()
-    expect(stack.canGoForward()).toBeFalsy()
-
-    stack.reset(ID_1)
-    expect(stack.current.value).toStrictEqual(ID_1)
-    stack.forward()
-    expect(stack.current.value).toStrictEqual(ID_1)
-    expect(stack.canGoBackward()).toBeFalsy()
-    expect(stack.canGoForward()).toBeFalsy()
-
-    stack.record(ID_2)
-    expect(stack.current.value).toStrictEqual(ID_2)
-    expect(stack.canGoBackward()).toBeTruthy()
-    expect(stack.canGoForward()).toBeFalsy()
-
-    stack.backward()
-    expect(stack.current.value).toStrictEqual(ID_1)
-    expect(stack.canGoBackward()).toBeFalsy()
-    expect(stack.canGoForward()).toBeTruthy()
-    stack.backward()
-    expect(stack.current.value).toStrictEqual(ID_1)
-    stack.forward()
-    expect(stack.current.value).toStrictEqual(ID_2)
-    expect(stack.canGoForward()).toBeFalsy()
-    stack.forward()
-    expect(stack.current.value).toStrictEqual(ID_2)
-
-    stack.backward()
-    expect(stack.current.value).toStrictEqual(ID_1)
-    stack.record(ID_3)
-    expect(stack.current.value).toStrictEqual(ID_3)
-    stack.forward()
-    expect(stack.current.value).toStrictEqual(ID_3)
-    stack.backward()
-    expect(stack.current.value).toStrictEqual(ID_1)
-  })
-}
