+++ app/gui2/src/stores/suggestionDatabase/documentation.ts
@@ -1,10 +1,9 @@
 import type { Group } from '@/stores/suggestionDatabase'
 import { findIndexOpt } from '@/util/data/array'
 import { isSome, type Opt } from '@/util/data/opt'
-import { unwrap } from '@/util/data/result'
 import { parseDocs, type Doc } from '@/util/docParser'
 import type { Icon } from '@/util/iconName'
-import { tryQualifiedName, type QualifiedName } from '@/util/qualifiedName'
+import { type QualifiedName } from '@/util/qualifiedName'
 
 export interface DocumentationData {
   documentation: Doc.Section[]
@@ -21,13 +20,15 @@ function isTagNamed(tag: string) {
   }
 }
 
-function tagValue(doc: Doc.Section[], tag: string): Opt<string> {
+/** @internal */
+export function tagValue(doc: Doc.Section[], tag: string): Opt<string> {
   const tagSection = doc.find(isTagNamed(tag))
   if (tagSection == null) return null
   return tagSection.Tag.body
 }
 
-function getGroupIndex(
+/** @internal */
+export function getGroupIndex(
   groupName: string,
   entryModule: QualifiedName,
   groups: Group[],
@@ -64,39 +65,4 @@ export function documentationData(
     isPrivate: isSome(tagValue(parsed, 'Private')),
     isUnstable: isSome(tagValue(parsed, 'Unstable')) || isSome(tagValue(parsed, 'Advanced')),
   }
-}
-
-if (import.meta.vitest) {
-  const { test, expect } = import.meta.vitest
-  const { initializeFFI } = await import('shared/ast/ffi')
-  await initializeFFI()
-
-  test.each([
-    ['ALIAS Bar', 'Bar'],
-    ['Some one section\n   But not tags here', null],
-    ['GROUP different tag', null],
-    ['PRIVATE\nGROUP Input\nALIAS Foo\n\nSeveral tags', 'Foo'],
-  ])('Getting tag from docs case %#.', (doc, expected) => {
-    const sections = parseDocs(doc)
-    expect(tagValue(sections, 'Alias')).toBe(expected)
-  })
-
-  const groups = [
-    { name: 'From Base', project: unwrap(tryQualifiedName('Standard.Base')) },
-    { name: 'Other', project: unwrap(tryQualifiedName('local.Project')) },
-    { name: 'Another', project: unwrap(tryQualifiedName('local.Project')) },
-  ]
-  test.each([
-    ['From Base', 'local.Project.Main', null],
-    ['From Base', 'Standard.Base', 0],
-    ['Standard.Base.From Base', 'local.Project.Main', 0],
-    ['Other', 'local.Project.Main', 1],
-    ['local.Project.Other', 'local.Project.Main', 1],
-    ['Other', 'local.Project.Some.Deep.Submodule', 1],
-    ['Another', 'local.Project.Main', 2],
-    ['Not Existing', 'local.Project.Main', null],
-  ])('Get group index case %#.', (name, definedIn, expected) => {
-    const definedInQn = unwrap(tryQualifiedName(definedIn))
-    expect(getGroupIndex(name, definedInQn, groups)).toBe(expected)
-  })
 }
