+++ app/gui2/src/util/ast/node.ts
@@ -31,30 +31,3 @@ export function primaryApplicationSubject(ast: Ast.Ast): Ast.AstId | undefined {
   if (!(ast instanceof Ast.Ident)) return
   return ast.id
 }
-
-if (import.meta.vitest) {
-  const { test, expect } = await import('vitest')
-  const { initializeFFI } = await import('shared/ast/ffi')
-  await initializeFFI()
-
-  test.each`
-    line                               | pattern      | rootSpan   | documentation
-    ${'2 + 2'}                         | ${undefined} | ${'2 + 2'} | ${undefined}
-    ${'foo = bar'}                     | ${'foo'}     | ${'bar'}   | ${undefined}
-    ${'## Documentation\n2 + 2'}       | ${undefined} | ${'2 + 2'} | ${'Documentation'}
-    ${'## Documentation\nfoo = 2 + 2'} | ${'foo'}     | ${'2 + 2'} | ${'Documentation'}
-  `('Node information from AST $line line', ({ line, pattern, rootSpan, documentation }) => {
-    const ast = Ast.Ast.parse(line)
-    const node = nodeFromAst(ast)
-    expect(node?.outerExprId).toBe(ast.id)
-    expect(node?.pattern?.code()).toBe(pattern)
-    expect(node?.rootSpan.code()).toBe(rootSpan)
-    expect(node?.documentation).toBe(documentation)
-  })
-
-  test.each(['## Documentation only'])("'%s' should not be a node", (line) => {
-    const ast = Ast.Ast.parse(line)
-    const node = nodeFromAst(ast)
-    expect(node).toBeUndefined()
-  })
-}
