+++ app/gui2/src/util/compare.ts
@@ -21,16 +21,3 @@ export function compareOpt(a: Opt<number>, b: Opt<number>, noneValueCmp: number 
     return isNone(a) ? noneValueCmp : -noneValueCmp
   }
 }
-
-if (import.meta.vitest) {
-  const { test, expect } = import.meta.vitest
-  test.each([
-    [1, 2, -1],
-    [2, 1, 1],
-    [1, 1, 0],
-    [null, 1, -1],
-    [1, null, 1],
-  ])('Compare %s with %s is %s', (a, b, expected) => {
-    expect(compareOpt(a, b)).toBe(expected)
-  })
-}
